# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/user/createUserWithExistingName": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Test New user creation",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/response.LoginResponse"
                        }
                    },
                    "404": {
                        "description": "JSON form error",
                        "schema": {
                            "$ref": "#/definitions/response.LoginResponse"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Super User login TODO?",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/response.LoginResponse"
                        }
                    },
                    "404": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.LoginResponse"
                        }
                    }
                }
            }
        },
        "/user/newLogin": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "New user creation",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/response.LoginResponse"
                        }
                    },
                    "404": {
                        "description": "JSON form error",
                        "schema": {
                            "$ref": "#/definitions/response.LoginResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.LoginResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}